Stop trying to deflate responses that redirect
Some servers (including YouTube) will send a redirect response with a
Content-Encoding header, but no body.  Since inflating an empty body
raises an exception, we should minimally work around this problem for
non-terminal redirect responses.

Note that this will still attempt to inflate a redirected response body
if HTTParty is not configured to follow redirects (which will raise
exceptions if the body is empty), and this does nothing to protect
against otherwise empty responses with a compressed Content-Encoding
(which may indicate a mis-configured server).