add Response#pretty_print
As discussed in #567, the changes introduced 5a78a14 resulted in change of
behavior when viewing the Response objects using command line version of the
HTTParty or tools like pry.

This commits adds minimal version of the #pretty_print method that behaves the
same as code before, when response was inheriting from BasicObject rather than
Object. This is because PP::ObjectMixin applies to Object, but it not
BasicObject.