Fix multipart to set its header even when other headers are provided
`@raw_request.initialize_http_header` resets the headers, so when
options contained :headers, it always erased the Content-Type header set
for multipart.

Now the Content-Type header is always set to "multipart/form-data" when
body is detected as multipart. This is IMHO not correct, but that's how
the original implementation was designed...